Added support for users and passwords as well as logging in.
Added support for individual credentials for end-points.

Changed GridFTP/FTP module to ignore PASV mode addresses, using only the
port, which is what many FTP clients do for security and reliability
reasons. The assumption is the command channel and the data channel
should always be on the same server.

Currently changing the way jobs are represented in the queue. The server
will now return much more "raw" ads, containing numerical representations
of various job statistics, allowing for clients to interpret and present
the data however appropriate. May add support for both raw and "cooked"
representation of queue in future.

Making progress on being able to store scheduler state to a file/stream
and recover it on startup. Currently configuration and users are
serializable.

Refactored the client code a little. Bug testing pending.